Adds SELECT test for eigenstates
Make sure SELECT applied to a target register that is set to an eigenstate
produces the correct state.

We diagonalze the Ising Hamiltonian and then perform <L|SELECT|L>
for the target register set to an eigenstate.  This results in
E_{l} / lambda.

Passing this check means unary iteration correctly applies the H to
the state.