Add 15-to-1 magic state factory (#693)
This PR is the first of a series of PRs that add T state distillation and improve the resource estimation workflow.

In this PR I implemented the 15-to-1 as it's presented in https://arxiv.org/abs/1905.06903. A few points of note:
1. The 4th rotation in Figure 3 gets moved to after the the 9th rotation.
2. I added comments to show which rotation is which.
3. all rotation which are supposed to be unitary in Figure 3 become channels that can either do the rotation $P\frac{\pi}{8}$ where $P$ is a pauli string or overshoot to $P\frac{5\pi}{8}$, $P\frac{-\pi}{8}$, or $P\frac{3\pi}{8}$.
4. Z and X errors are created as [cirq.AsymmetricDepolarizingChannel](https://quantumai.google/reference/python/cirq/asymmetric_depolarize) channels using the `storage_error` method.
5. The probabilities in each of the channels are the same as in the [supplementary materials](https://github.com/litinski/magicstates/blob/79d99a5f21e63f992f63e644ea43a665944b4bb5/Python/onelevel15to1.py#L8) of https://arxiv.org/abs/1905.06903.
6. The `n_cycles` method in magic factories should take a physical error parameter. the CCZ model that we have https://arxiv.org/abs/1812.01238 assumes physical error rate of $1e-3$ (`Assuming a physical gate error rate of 10^−3, our CCZ factory can produce ...`)

---

This PR reintroduces MagicCount dataclass to be used exclusively with magic factories. the reason for this is that although AlgorithmSummary is a superset of MagicCount, it becomes unclear what to do when AlgorithmSummary has other counts (e.g. rotations). one way would be to error out but I think a seprate class should make things cleaner.