Create linear half comparison bloqs (#1408)
This PR creates all versions of half a comparison operation in $n$ toffoli complexity. The uncomputation of these bloqs has zero cost.

The half comparason operation is needed in the modular inversion bloq.

I also modify the OutOfPlaceAdder to allow it to not compute the extra And when it's not needed

---

This PR also shows how we can implement all versions of comparison by implementing just one. for example creating a half greater than bloq that use logarithmic depth is enough to implement all the others.