Fix test_qft_text_book following #1455 (#1574)
gh-1455 included a change to test_qft_text_book_quick to respect the
new, native CRz bloqs. I must have missed porting these changes
to the @pytest.mark.slow version of this unit test, and the nightly
build failed. This ports the changes to the slow test to fix the
nightly build.

Co-authored-by: Noureldin <noureldinyosri@google.com>