Do minor updates to dot files and add license headers (#1598)
* Add license header

* Update to current common version

Changes:
- Add license header
- Add line length setting
- Reduce allowed raw HTML elements

This version matches the current common version of the file I've been
adding to the Quantumlib repos.

* Update to current common version

Changes:
- Add license header
- Add more settings
- Adjust settings to match (apparent) Qualtran project conventions

Except for the changed settings, this version of the file matches the
current common version of the file I've been adding to the Quantumlib
repos.

* Add back max_line_length setting

After more discussion, the conclusion is that despite the issues with
`max_line_length`, it is handled as expected in the editors that many
people have been using.