Add CITATION.cff file (#1597)
* Add CITATION.cff file

* Add more comments & edit mildly

This adds explanatory comments, and very slightly edits formatting and
the keyword list.

* Fix title for the home page link

* Edit confusing comment

* Fix incorrect preferred-citation "type" field value

* Fix more fields in preferred-citation

Man, sometimes the CITATION.cff limitations are so annoying. This time
it's because "volume" has to be a plain integer.

* Add "Citation" section to README

* Add automation to update CITATION.cff upon releases

This GitHub Actions workflow will update the version and date fields
in the CITATION.cff file when a new release is done in GitHub.

* Do tiny edits to comments in CITATION.cff

* Add more data to the preferred citation

* Remove archivePrefix & primaryClass from BibTex record

They're nonstandard and don't seem to be used much.

* Do not add citation-updater.yaml at this time

Needs more work.