Expand and embellish README file (#1601)
* Add "Citation" section to README

* Remove trailing whitespace

In Markdown, two trailing spaces on a line are a shorthand way to
indicate a forced line break. This is prone to causing problems
accidentally, so as a safety precaution, it's a good idea to remove
any whitespace at the ends of lines.

* Add required Google disclaimer & also contact info

Google requires the text of this disclaimer in project README files
for open-source projects.

* Add QAI logo to bottom of README file

It's pretty, and it's good for branding.

* Add Qualtran logo files

* Edit, expand, and embellish README file

Changes:

* Add required Google disclaimer
* Add new "Usage", "News" & "Contact" sections
* Move some text around
* Use GitHub Markdown syntax for warning boxes
* Add a TOC line near the top, for rapid access
* Add logos and badges, for some extra visual flair

* Add HTML anchor for "how-to-cite"

The section "How to cite" appears in a number of other Quantumlib
READMEs, and URLs referencing #how-to-cite exist in some documens, so
for maximum compatibility, let's add an additional anchor for links.

* Change smallcaps to normal font

* Change section News → Community & add more content

* Remove mention about the GUI

Per [review comment by
@mpharrigan](https://github.com/quantumlib/Qualtran/pull/1601#discussion_r2017533418),
removing the mention of the physical resources estimation GUI.

* Following up on suggestion to put Qualtran title in comment

Following up on a suggestion in a [review comment by
@mpharrigan](https://github.com/quantumlib/Qualtran/pull/1601#discussion_r2017530092)
to put the Qualtran title in comment at the top of the file.

* Replace separate dark/light logo files with one combined SVG

Using a trick from https://stackoverflow.com/a/67190894/743730, we can
create a single SVG that will change text colors depending on whether
the user is using dark mode or light mode on GitHub.

* Switch to new single SVG logo file

* Remove statement about Qualtran being an abbreviation

Because it isn't.