Update message text
It's unfortunate that Zenodo puts the `message:` field text
prominently in a yellow box on the Zenodo archive page for the
software. IMHO, this runs counter to the definition of the field in
the CITATION.cff format specification ("A message to the human reader
**of the CITATION.cff file** to let them know what to do with the
citation metadata", emphasis added). It makes it difficult to come up
with text that can serve both the role defined by the CITATION.cff
spec and Zenodo's way of displaying it.