Doc system notebook execution: WriteIfDifferent. Misc doc fixes (#1648)
Feature: `jupyter_autogen.py` will only re-write `*.ipynb` files if
there's been a change.
The tool to execute the notebooks is already to set up to only
re-execute notebooks if they're out of date. But every time you re-run
the first autogen script, it will re-up the modification date and
everything will always be re-evaluated, which is expensive.

Misc fixes
- Use a distinct extension for `autotoc.rst.inc`, so sphinx doesn't
treat it as a page and emit warnings about it
 - include missing TOC entries for notebooks
 - fix flaky notebook assertion by using more idiomatic sympy methods
 - grammar and arg docstring fixes in gf_poly