Fix additional bug in ECAdd (#1666)
Team from PsiQuantum reached out and pointed out another bug in the
circuit nearly equivalent to one of the other ones that I found/fixed.

When [(a, b) = (0, 0) AND x = 0] OR [(x, y) = (0, 0) AND a = 0], the f_1
flag is not cleared. I added the same fix to this situation that I did
in the similar scenario when the b/y values are all 0 for the f_2 flag.
This cost 2 3n-controlled toffolis.

---------

Co-authored-by: Matthew Harrigan <mpharrigan@google.com>