Fix recursive phased classical simulation (#1731)
The classical simulators should recurse into bloq decompositions. The
phased classical simulator would use the ordinary classical simulator
when recursing since that was "baked into" the `Bloq` method -- and it
would throw an error even if it should have been able to simulate the
(sub-)circuit.

This changes the responsibility of dealing with the possible recursion
to the simulator classes themselves.